@@ -3,19 +3,22 @@
   "author": "n.m.batrakov",
   "license": "MIT",
   "scripts": {
-    "install": "lerna bootstrap",
+    "install": "lerna bootstrap --hoist",
     "build": "lerna run build",
     "test": "jest",
     "lint": "tslint --project .",
-    "publish": "lerna run build && lerna publish"
+    "publish": "lerna run build && lerna publish",
+    "lerna": "lerna"
   },
   "devDependencies": {
-    "@types/jest": "^24.0.23",
-    "jest": "^24.9.0",
-    "lerna": "^3.18.4",
-    "ts-jest": "^24.1.0",
-    "tslint": "^5.20.1",
+    "@types/jest": "^25.1.4",
+    "@types/node-fetch": "^2.0.0",
+    "jest": "^25.2.3",
+    "lerna": "^3.20.2",
+    "node-fetch": "^2.0.0",
+    "ts-jest": "^25.2.1",
+    "tslint": "^6.1.0",
     "tslint-config-airbnb": "^5.11.2",
-    "typescript": "^3.7.2"
+    "typescript": "^3.8.3"
   }
 }
