@@ -5,66 +5,66 @@ import { createPipeline, Pipes } from '../pipe'
 type Etl = (config: any) => OperatorFunction<void, any>
 
 describe('createPipeline', () => {
-    const sourceA: Etl = () => s => s.pipe(
+    let sourceA: Etl = () => s => s.pipe(
         mergeMap(() => interval(10).pipe(take(3))),
         map(x => `A${x}`),
     )
 
-    const sourceB: Etl = () => s => s.pipe(
+    let sourceB: Etl = () => s => s.pipe(
         mergeMap(() => interval(12).pipe(take(3))),
         map(x => `B${x}`),
     )
 
     it('sequential', async () => {
-        const observable = testSequential({ A: sourceA, B: sourceB }, ['A', 'B'])
+        let observable = testSequential({ A: sourceA, B: sourceB }, ['A', 'B'])
 
-        const expected = ['A0', 'A1', 'A2', 'B0', 'B1', 'B2']
-        const actual = await observable.pipe(toArray()).toPromise()
+        let expected = ['A0', 'A1', 'A2', 'B0', 'B1', 'B2']
+        let actual = await observable.pipe(toArray()).toPromise()
 
         expect(actual).toEqual(expected)
     })
 
     it('concurrent', async () => {
-        const observable = testConcurrent({ A: sourceA, B: sourceB }, ['A', 'B'])
+        let observable = testConcurrent({ A: sourceA, B: sourceB }, ['A', 'B'])
 
-        const expected = ['A0', 'B0', 'A1', 'B1', 'A2', 'B2']
-        const actual = await observable.pipe(toArray()).toPromise()
+        let expected = ['A0', 'B0', 'A1', 'B1', 'A2', 'B2']
+        let actual = await observable.pipe(toArray()).toPromise()
 
         expect(actual).toEqual(expected)
     })
 
     it('sequential respect order', async () => {
-        const observable = testSequential({ A: sourceA, B: sourceB }, ['B', 'A'])
+        let observable = testSequential({ A: sourceA, B: sourceB }, ['B', 'A'])
 
-        const expected = ['B0', 'B1', 'B2', 'A0', 'A1', 'A2']
-        const actual = await observable.pipe(toArray()).toPromise()
+        let expected = ['B0', 'B1', 'B2', 'A0', 'A1', 'A2']
+        let actual = await observable.pipe(toArray()).toPromise()
 
         expect(actual).toEqual(expected)
     })
 
     it('sequential filter', async () => {
-        const observable = testSequential({ A: sourceA, B: sourceB }, ['B'])
+        let observable = testSequential({ A: sourceA, B: sourceB }, ['B'])
 
-        const expected = ['B0', 'B1', 'B2']
-        const actual = await observable.pipe(toArray()).toPromise()
+        let expected = ['B0', 'B1', 'B2']
+        let actual = await observable.pipe(toArray()).toPromise()
 
         expect(actual).toEqual(expected)
     })
 
     it('concurrent filter', async () => {
-        const observable = testConcurrent({ A: sourceA, B: sourceB }, ['B'])
+        let observable = testConcurrent({ A: sourceA, B: sourceB }, ['B'])
 
-        const expected = ['B0', 'B1', 'B2']
-        const actual = await observable.pipe(toArray()).toPromise()
+        let expected = ['B0', 'B1', 'B2']
+        let actual = await observable.pipe(toArray()).toPromise()
 
         expect(actual).toEqual(expected)
     })
 
     it('run all on when scripts empty', async () => {
-        const observable = testSequential({ A: sourceA, B: sourceB }, [])
+        let observable = testSequential({ A: sourceA, B: sourceB }, [])
 
-        const expected = ['A0', 'A1', 'A2', 'B0', 'B1', 'B2']
-        const actual = await observable.pipe(toArray()).toPromise()
+        let expected = ['A0', 'A1', 'A2', 'B0', 'B1', 'B2']
+        let actual = await observable.pipe(toArray()).toPromise()
 
         expect(actual).toEqual(expected)
     })
@@ -72,17 +72,17 @@ describe('createPipeline', () => {
 
 
 function testSequential<T = any>(pipes: { [name: string]: Etl }, filter: string[]): Observable<T> {
-    const pipeline = mapObjPipesToArray(pipes)
+    let pipeline = mapObjPipesToArray(pipes)
 
-    const operator = createPipeline(pipeline, filter, false)({})
+    let operator = createPipeline(pipeline, filter, false)({})
 
     return of(undefined).pipe(operator)
 }
 
 function testConcurrent<T = any>(pipes: { [name: string]: Etl }, filter: string[]): Observable<T> {
-    const pipeline = mapObjPipesToArray(pipes)
+    let pipeline = mapObjPipesToArray(pipes)
 
-    const operator = createPipeline(pipeline, filter, true)({})
+    let operator = createPipeline(pipeline, filter, true)({})
 
     return of(undefined).pipe(operator)
 }
