@@ -1,27 +1,23 @@
 import { OperatorFunction, of, merge, concat, Observable } from 'rxjs'
 import { notNullOrUndefined } from '../utils'
 
-export type EtlPipe = (config: any) => OperatorFunction<void, any>
+export type EtlPipe = (config: any) => OperatorFunction<any, any>
 
 export type Pipes = Array<{
     name?: string,
     pipe: EtlPipe,
 }>
 
-export function runPipeline(pipeline: OperatorFunction<void, any>) {
-    return pipeline(of(undefined)).toPromise()
-}
-
 export function createPipeline(
     pipes: Pipes,
     scripts: string[],
     concurrent: boolean,
-): (config: any) => OperatorFunction<void, any> {
-    const filtered = scripts.length === 0
+): EtlPipe {
+    let filtered = scripts.length === 0
         ? pipes
         : scripts.map(name => pipes.find(x => x.name === name)).filter(notNullOrUndefined)
 
-    const combine = concurrent ? merge : concat
+    let combine = concurrent ? merge : concat
 
     return (config: any) =>
         (stream: Observable<void>) =>
@@ -29,7 +25,7 @@ export function createPipeline(
 }
 
 export function isScriptsValid(pipes: Pipes, scripts: string[]) {
-    const pipesKeys = pipes.map(x => x.name).filter(notNullOrUndefined)
+    let pipesKeys = pipes.map(x => x.name).filter(notNullOrUndefined)
 
     if (pipesKeys.length === 0) {
         console.error('Error: Unable to run ETL - there are no named scritps')
