@@ -1,4 +1,4 @@
-export * from './http'
+export * from './promisify'
 export * from './logging'
 export * from './array'
 
@@ -11,3 +11,10 @@ export function isNullOrUndefined<T>(value: T | null | undefined): value is null
 export function notNullOrUndefined<T>(value: T | null |undefined): value is T {
     return value !== undefined && value !== null
 }
+
+export const throwError = (e: string | Error) => { throw e }
+
+export const not = <T>(predicate: (x: T) => boolean) => (x: T) => !predicate(x)
+
+export type Type = 'string' | 'number' | 'bigint' | 'boolean' | 'symbol' | 'undefined' | 'object' | 'function'
+export const ofType = <T = any>(type: Type) => (x: T): x is T => typeof x === type
